Point SwiftLint fork to branch with smaller diff
I'm using this repo as an example to show the fix for the remote config
to open a PR against the original feature on SwiftLint.

I had forgotten I had rebased my fork on SwiftLint's master, which would
make opening a PR against the feature branch hard due to a longer diff.

Switching to a branch with just the fix on top of the original feature
branch so that I can show my fix is effective with a smaller diff.